Check if Burrow is already running
When Burrow starts, it checks to see if the PID file exists. If it does,
it kills itself and returns 1. If it doesn't it starts itself and writes
the file. This commit now checks to see if there is a file and if there
is one, it reads it and checks to see if the PID in the file exists. If
it does, it kills itself. If it doesn't it deletes the file and starts
itself writing a new file.